services/email-notifier/requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
email-validator>=2.0.0
jinja2>=3.1.0
pytz>=2023.3
pydantic>=2.0.0
pgvector>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

